requirements.txt
Cerberus==0.9.2
Eve==0.6.1
Events==0.2.1
Flask==0.10.1
Flask-PyMongo==0.3.1
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
py-bcrypt==0.4
pymongo==2.9.2
simplejson==3.8.2
Werkzeug==0.10.4
wheel==0.24.0
